fdr.ma {unknown}	R Documentation

fdr.ma

Description

This function takes normalized expression data array, experimental design and computes adjusted p-values It returns the fdr adjusted p-values and plots, according to the methods described in (Reiner, Yekutieli and Benjamini 2002) save.

Usage

fdr.ma(exp.arr,design,p.method="resampling",fdr.adj="BH-LSU",out="adj",plot="t",perms.num=100)

Arguments

exp.arr 	 A data frame or matrix, with m rows corresponding to the gene expression level of an observation and n columns to the observations. In the case of gene expression data, rows correspond to genes and columns to mRNA samples. The data can be read using 'read.table'. The matrix is as the X matrix in multtest package.

design 	 A vector of charachters corresponding to observation (column) class labels. For k classes, the labels must be k different charachters. The vector is similar to classlabel vector in multtest package.

p.method 	 A character string specifying whether resampling mehtod may be used. If 'p.method="resampling"', resampling method is used. If 'p.method="t.test"', t.test is computed without resampling.

fdr.adj 	 A character string specifying which type of p-value adjustment method may be used. If 'fdr.adj="BH-LSU"', Benjamini-Hochberg Linear Step Up porcedure is used. If 'fdr.adj="point.est"', point estimation procedure is used.(p.method value must be "resampling") If 'fdr.adj="upper.est"', upper estimation procedure is used.(p.method value must be "resampling") If 'fdr.adj="adaptive"', two-stage adaptive procedure is used.

out 	 A character string specifying what value may be shown on the Y axis of the plot. If 'out="adj"', the y axis values are the adjusted p-values. If 'out="pvalue"', the y axis values are the original p-values. If 'out="both"', both the original and adjusted p-values are drawn in two figures.

plot 	 A character string specifying what value may be shown on the x axis of the plot. If 'plot="TRUE"', the x axis values are the t-test values. If 'plot="diag"', the x axis values are the rank of the original p-value of the hypothesis.

perms.num 	 An integer specifying the number of permutations that would be used.

Value

A data frame with components

adj 	 adjusted p-values.

dif 	 the differences between the means of the two groups.

res.pvalues 	 resampling p-values. (in case resampling was requested)

Author(s)

Anat Reiner Daniel Yekutieli Yoav Benjamini Effi Kenigsberg

References

Reiner A, Yekutieli D, Benjamini Y: Identifying differentially expressed genes using false discovery rate controlling procedures. Bioinformatics 19:368-375, 2003 Benjamini, Y., Krieger, A.M.,Yekutieli, D. (2001) “Two Staged Linear Step Up FDR Controlling Procedure”, Technical Report Department of Statistics and O.R., Tel Aviv University.

See Also

fdr.gui.

Examples


exp.arr<<-as.matrix(read.table("exp.arr"))
design<-c(0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1)
output<-fdr.ma(exp.arr,design,p.method="t.test",fdr.adj="BH-LSU",out="adj",plot="TRUE")

[Package Index]